Sam suggested renaming `model.py` to `components.py`
