+++ ionerdss/model/platonic_solids_model.py
@@ -3,7 +3,7 @@
 This module defines the `PlantonicSolidsModel` class, inheriting from the `Model` class, which is used to generate NERDSS molecule types and reactions, and corresponding files for platonic solid.
 """
 
-from .model import Model
+from .components import Model
 
 class PlantonicSolidsModel(Model):
     """"A class for generating NERDSS molecule types and reactions, and corresponding files for platonic solid.
